requirements.txt
click==8.1.3
Django==4.1.6
djangorestframework==3.14.0
openpyxl==3.1.0
packaging==23.0
pymemcache==4.0.0
python-dateutil==2.8.2
pytz==2022.7.1
